backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==0.28.1
pinecone-client==2.2.4
langchain==0.0.348
gTTS==2.4.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.1 